requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
flask-cors>=4.0.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
anthropic>=0.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
